+++ src/hooks/announcements/useHandleAnnouncements.ts
@@ -1,4 +1,4 @@
-import { DisplayMediaType } from "@/src/components/announcements/AddMedia";
+import { DisplayMediaType } from "@/src/components/announcements/media/AddMedia";
 import submitPost from "@/src/components/announcements/submitPostData";
 import {
   getDefaultAnnouncement,
